[HAPI-184] Create the CRUD class instance (#200)
* [HAPI-184] Create the CRUD class instance

## [HAPI-184] Description
Created a class to manage the CRUD operations

* [HAPI-184] Fixing crud filename

* [HAPI-184] Fixing CRUD filename

* [HAPI-184] Fixing issue with merge
