--- src/services/database/index.js
@@ -1,7 +1,0 @@
-const init = require('./init');
-const crud = require('./crud');
-
-module.exports = {
-    init,
-    crud
-};
